LayoutTests/imported/w3c/web-platform-tests/domparsing/innerhtml-05-expected.txt

CONSOLE MESSAGE: line 1380: Blocked a frame with origin "http://localhost:8800" from accessing a frame with origin "null".  The frame requesting access has a protocol of "http", the frame being accessed has a protocol of "data". Protocols must match.



FAIL innerHTML in XHTML undefined is not an object (evaluating 'window[0].document.documentElement')
